ReservedBook Tests + Validation

+++ LibraryWebParent/LibraryBackend/LibraryReader/src/main/java/gb/library/reader/controllers/ReservedBookController.java
@@ -4,8 +4,11 @@ import gb.library.common.entities.ReservedBook;
 import gb.library.reader.converters.ReservedBookConverter;
 import gb.library.reader.dtos.ReservedBookReaderDto;
 import gb.library.reader.services.ReservedBooksService;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -23,12 +26,16 @@ public class ReservedBookController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteFromReserved(@PathVariable int id) {
+    public ResponseEntity<?> deleteFromReserved(@PathVariable  int id) {
+        if (id < 1)
+            return ResponseEntity.badRequest().build();
+
         reservedBooksService.delete(id);
+        return ResponseEntity.noContent().build();
     }
 
     @PostMapping("/create")
-    public ReservedBookReaderDto reserveBook(@RequestBody ReservedBookReaderDto reservedBookDto) {
+    public ReservedBookReaderDto reserveBook(@RequestBody @Valid ReservedBookReaderDto reservedBookDto) {
         ReservedBook book = reservedBooksService.create(converter.DtoToEntity(reservedBookDto));
         return converter.entityToDto(book);
     }
